Initial stab at a Clojure plugin
Sadly, this does not work, because Clojure requires the current
thread's context class loader to be set to a loader that finds
Clojure. ImageJ 1.x does not do that. The only way to do it would be
to wrap the Clojure code into a real Java class which kind of makes
no sense at all.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>