POM: remove redundant configuration blocks
Let's not make it _too_ easy for people to override Maven's very
sensible defaults with something less sensible -- even if there are
times when you need to do it. Those times should be the exception
to the rule, and those people need to learn a little bit more about
Maven in that case.