Set plugins.dir correctly for paths that spaces
In Process_Pixels, `plugins.dir` is set manually to the outpath of the
compiled .class files by querying `class.getResource`.
However, this does not work if the path contains whitespaces that are
escaped by `getResource`.

This PR fixes this issue for spaces by replacing their escape code
(`%20`) by the original character.